revise tests
