+++ tests/pdk/finfet_pdk/test_placement_grid.py
@@ -4,11 +4,21 @@ import pytest
 import textwrap
 from .utils import get_test_id, build_example, run_example
 from . import circuits
+from align.pdk.finfet import MOSGenerator
 
 
 @pytest.fixture(autouse=True)
 def place_on_grid(monkeypatch):
     monkeypatch.setenv('PLACE_ON_GRID', 't')
+
+
+@pytest.mark.parametrize('vt', ['NMOS', 'PMOS'])
+def test_check_constraints(vt):
+    c = MOSGenerator()
+    ports = {'S': [('M1', 'S')], 'D': [('M1', 'D')], 'G': [('M1', 'G')]}
+    parameters = {'M': 1, 'NFIN': 4, 'real_inst_type': vt, 'NF': 2}
+    c.addNMOSArray(1, 1, 0, None, ports, **parameters)
+    assert c.metadata["constraints"][0]["constraint"] == "place_on_grid"
 
 
 def test_scalings():
